@@ -3,17 +3,23 @@
   "version": "0.0.0",
   "private": true,
   "dependencies": {
-    "@angular/animations": "^12.1.0",
-    "@angular/common": "^12.1.0",
-    "@angular/compiler": "^12.1.0",
-    "@angular/core": "^12.1.0",
-    "@angular/forms": "^12.1.0",
-    "@angular/platform-browser": "^12.1.0",
-    "@angular/platform-browser-dynamic": "^12.1.0",
-    "@angular/router": "^12.1.0",
+    "@angular/animations": "^12.2.4",
+    "@angular/cdk": "^12.2.4",
+    "@angular/common": "^12.2.4",
+    "@angular/compiler": "^12.2.4",
+    "@angular/core": "^12.2.4",
+    "@angular/forms": "^12.2.4",
+    "@angular/material": "^12.2.4",
+    "@angular/platform-browser": "^12.2.4",
+    "@angular/platform-browser-dynamic": "^12.2.4",
+    "@angular/router": "^12.2.4",
+    "@asymmetrik/ngx-leaflet": "^8.1.0",
+    "leaflet": "^1.7.1",
+    "ngx-toastr": "^14.1.0",
     "rxjs": "^7.1.0",
+    "rxjs-compat": "^6.6.7",
     "tslib": "^2.3.0",
-    "zone.js": "^0.11.4"
+    "zone.js": "~0.11.4"
   },
   "scripts": {
     "ng": "ng",
@@ -24,15 +30,16 @@
     "e2e": "ng e2e"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1100.7",
-    "@angular/cli": "~11.0.4",
-    "@angular/compiler-cli": "^11.0.9",
+    "@angular-devkit/build-angular": "^12.2.4",
+    "@angular/cli": "~12.2.4",
+    "@angular/compiler-cli": "^12.2.4",
     "@types/jasmine": "~3.6.0",
+    "@types/leaflet": "^1.7.5",
     "@types/node": "^12.11.1",
     "codelyzer": "^6.0.0",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.1.0",
+    "karma": "~6.3.4",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
@@ -40,6 +47,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.0.2"
+    "typescript": "~4.3.5"
   }
 }
