@@ -4,11 +4,6 @@ import { FormsModule } from '@angular/forms';
 import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 
 import { AppComponent } from './app.component';
-import { ProductListComponent } from './product/product-list.component';
-import { ProductDetailComponent } from './product/product-detail.component';
-import { ProductService } from './product/product.service';
-import { CategoryService } from './category/category.service';
-import { CategoryListComponent } from './category/category-list.component';
 import { AppRoutingModule } from './app-routing.module';
 import { DashboardComponent } from './dashboard/dashboard.component';
 import { NotifyComponent } from './notify/notify.component';
@@ -22,8 +17,7 @@ import { ErrorComponent } from './core/error.component';
 import { RegisterComponent } from './register/register.component';
 import { GroupComponent } from './group/group.component';
 import { HttpConfigInterceptor } from './core/error/httpconfig.interceptor';
-import { SignalRComponent } from './notiy/signal-r.component';
-import { SignalRService } from './notiy/signal-r.service';
+
 import { RtgpsComponent } from './notify/rtgps.component';
 import { RtgpsService } from './notify/rtgps.service';
 import { LeafletModule } from '@asymmetrik/ngx-leaflet';
@@ -40,16 +34,13 @@ import { DownloadComponent } from './download/download/download.component';
 @NgModule({
   declarations: [
     AppComponent,
-    ProductListComponent,
-    ProductDetailComponent,
-    CategoryListComponent,
     DashboardComponent,
     NotifyComponent,
     LoginComponent,
     ErrorComponent,
     RegisterComponent,
     GroupComponent,
-    SignalRComponent,
+
     RtgpsComponent,
     TestComponent,
     LoaderComponent,
@@ -67,15 +58,12 @@ import { DownloadComponent } from './download/download/download.component';
       positionClass: 'toast-top-right',
       preventDuplicates: false
     }),
-    LeafletModule.forRoot(),
+    LeafletModule,
     MatProgressSpinnerModule
   ],
   providers: [
-    ProductService,
-    CategoryService,
     SecurityService,
-    SignalRService,    
-    RtgpsService,
+     RtgpsService,
     { provide: ErrorHandler, useClass: GlobalErrorHandler },
     { provide: HTTP_INTERCEPTORS, useClass: HttpConfigInterceptor, multi: true }
   ],
